partial commit, not to be used

+++ docker/veritas/build-docker.sh
@@ -1,14 +1,19 @@
 #!/bin/bash
 
-#if ! [ -d "bin" ]; then
+if ! [ -d "../../bin" ]; then
+    echo "Please build the binaries first! (cd ../../scripts; ./build_binaries.sh)"
+    exit 1
+fi
+
 rm -rd bin
 cp -r ../../bin .	
-#fi
+
 if ! [ -f "id_rsa.pub" ]; then
-    if ! [ -f "~/.ssh/id_rsa.pub" ]; then
-        echo "You do not have a public SSH key. Please generate one with 'ssh-keygen'!"
+    if ! [ -f "$HOME/.ssh/id_rsa.pub" ]; then
+        echo "You do not have a public SSH key. Please generate one! (ssh-keygen)"
         exit 1
     fi
-    cp ~/.ssh/id_rsa.pub .
+    cp $HOME/.ssh/id_rsa.pub .
 fi
+
 docker build -f Dockerfile -t veritas .
